Move TopicLogger out into separate enterprise package
The TopicLogger is a mechanism for forwarding Redpanda Connect logs
directly into a kafka topic. It was initially implemented within the
kafka package since it reused some of the libraries within there for
configuration and so on.

This means that an import of the kafka components was importing an
enterprise licensed feature. In practice this is not a concern for
FOSS users as it's impossible to activate this code simply by importing
the package. However, for clarity it's better for this feature to
exist in a separate package altogether.