Pinecone output
Support an output to pinecone vector database. We support three
different kind of operations: upsert (default), update, delete.

There are mappings for metadata and vectors in Pinecone, as well as
interpolations for ID and namespaces.

The tests here are rather slim as there isn't a ton of logic here, it's
just hooking stuff up, and there is no way to write an isolated
integration test (pinecone is only a hosted service).