snowpipe/int128: add from float support (with scaling and precision)
So this commit adds support for scale!=0 in Int128 values, this is
important for supporting all possible numeric values in snowflake,
but this is a narly bit of math here. All the code here is adapted from
the apache arrow project (licensing is properly attributed) with some
light fixes around edge cases.

I will attempt to upstream this and our bug fixes/perf improvements to
the arrow project so we don't have to maintain this code forever. If
that happens we can take a dependency on apache arrow go and drop this
int128 package in our tree.