snowpipe: fix type narrowing issue
Talking with the snowflake team we don't have to narrow types if we
omit the `sfVer` key in the metadata. So remove that and simplify the
type writing bits.

NOTE: we also simplify the timestamp conversion stuff to only support
standard RFC8339Nano format. If someone wants to support another format
they will need to parse the date/time/timestamp in blobl.