snowflake: optimize row writing
Remove a map assignment in the hot path of converting data, the loop
here seems to be most of the cost, so use an array + offset lookup
approach for matching values to the transformer/converter.

We still *have* to have a map lookup due to the user giving us data by
name (I don't think we want to encourage people to get the ordinals
correct).

If this loop is *still* slow we could consider parallizing the
conversion here with multiple goroutines. Stats would need to be
threadsafe or we'd need a converter per goroutine and merge the stats
at the end (probably not too bad).