snowflake: detach stats from transformers
Let stats be an article of constructing the row group, instead of a
persistent object that is shared and reset. There are no allocations
saved by reusing stats and now we can parallelize row group
construction, as there is no shared mutable state in the transformers
outside of the typed buffer.