snowflake: schema evolution
Enable schema evolution using a custom mapping function, to allow
customers to customize the evolution method. Schema evolution needs to
be a global pause so that we can reopen channels to get new schemas.

I'm not in love with how errors are being hijacked here but I have not
came up with a better idea in the short time working on.