pgcdc: remove extra roundtrip for LSNs
This for silly reasons goes from string->int->string for LSNs when
storing in the checkpointer. The uint64 should use less memory than
string but I don't think we really care about that level of
optimization, so just keep it as a string everywhere and simplify the
code.

I'll buy you lunch if we end up needing that extra memory ever :)