pgcdc: move all pg conn usage to a single goroutine
s.pgConn is not safe to use in multiple goroutines, so move all it's
usage to the single goroutine that is emitting changes.

This also allows us to do the annoying "remapping LSN for last message
in a txn" locally which simplifies a lot of things and also should speed
stuff up because we no longer need to wait for a txn to be acked before
moving on to the next one. See the comments in the code