inputs/redpanda: add fetch_max_wait option
kgo.FetchMaxWait is a config option supported by franz-go. It makes it
possible to use kgo.FetchMinBytes, but have a rather low max wait time
to fill the batch. This makes it possible to force the broker to send
big batches if possible, but still wait only for a short time if there's
not enough data.

This is especially important with the redpanda input, as it's using
ordered franz-go. It will only send batches, if the previous batch with
the partition has been consumed. If the broker keeps sending very small
batches, e.g. size 1, it's likely to stall batched outputs. I could
reproduce locally by using a producer that sends lots of batches of size
1.

Using kgo.FetchMinBytes in combination with kgo.FetchMaxWait can solve this
problem.
But in any case, it is a useful tuning knob offered by franz-go, but
also the standard Java client.