pgcdc: fix snapshotting tx bug
Before this we were creating transactions on connection pools, so
subsequent transactions didn't even get the right connection always.

This is a nasty consistency bug with the initial snapshot. We also never
seemed to close the initial transaction in postgres, so it was holding
locks and mucking stuff up...

Fix all this by using the sql package correctly and explicitly using
transaction support from the sql package.