mongo/cdc: initial implementation
The initial implementation of MongoDB CDC is not too unlike debezium.

Generally the idea is:

* Get the latest committed message in the oplog
* Read all specified collections in parallel
* Stream changes from the previously recorded oplog position

this allows us to get everything, but we may not have an exactly
consistent view when replaying changes that happened during the
snapshot phase, but it will be eventually consistent. I believe
with more recent MongoDB versions we can ensure more consistency
using snapshot read isolation, but I'm not sure there is a large
benefit to using that. As the issue is you can't read in parallel
with snapshots, since snapshots are limited to a single transaction
as far as I can tell.

This also requires that MongoDB is running in "replication set" mode,
which is the recommended mode for production usage, but requires some
tricks to get a single node docker setup working correctly, which is
why I used test containers instead of dockertest for the integration
tests.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>