Update some dependencies (#3202)
* Update some deps

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Update test following goawk bug fix

Basically, `{ }` should not be equivalent to `{ print $0 }`, so
then the message will remain unchanged.

Details: https://github.com/benhoyt/goawk/pull/230

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Fix OpenAI code

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Fix Sentry integration tests

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Ignore deprecation warning for jhump/protoreflect

We'll have to replace this with github.com/bufbuild/protocompile.

It's probably not a lot of work, but it needs some testing and we
have some open issues around protobuf handling.

Signed-off-by: Mihai Todor <todormihai@gmail.com>

---------

Signed-off-by: Mihai Todor <todormihai@gmail.com>