git: at least once delivery when checkpointing
We need to wait to checkpoint until all messages in a commit has been
ack'd. So we do that in a crude way by not processing any more data
until everything has been acked from the commit, then checkpointing.

This is less than ideal in that ideally we'd like to continue the
pulling/processing if there is just a couple messages in flight, but
this at least allows us to provide some kind of backpressure if the
pipeline is really slow compared to the pull interval.

Also cleanup error handling a bit.