Back off unrecoverable conn errors (#3529)
* Back off unrecoverable conn errors

Currently we silently retry kafka broker connections indefinitely
even when the error is something unrecoverable, such as a SASL
credentials issue.

Under these circumstances the connections remained open and our
attempts were being made so frequently that it was causing issues
at the broker level.

As a fix this commit attempts to detect unrecoverable connection
errors and, when detected, forces a client closure as well as a
full minute of back off before another attempt is made.

* Simplify implementation as we can just use ping